requirement.txt
mysql-connector-python
pandas
openpyxl
pygad